+++ demisto_sdk/commands/common/tests/integration_test.py
@@ -519,3 +519,20 @@ class TestIsFeedParamsExist:
             self.validator.current_file['configuration'].append(t)
 
         assert self.validator.all_feed_params_exist() is False, 'all_feed_params_exist() returns True instead False'
+
+    NO_HIDDEN = {"configuration": [{"id": "new", "name": "new", "display": "test"}, {"d": "123", "n": "s", "r": True}]}
+    HIDDEN_FALSE = {"configuration": [{"id": "n", "hidden": False}, {"display": "123", "name": "serer"}]}
+    HIDDEN_TRUE = {"configuration": [{"id": "n", "n": "n"}, {"display": "123", "required": "false", "hidden": True}]}
+    HIDDEN_TRUE_AND_FALSE = {"configuration": [{"id": "n", "hidden": False}, {"ty": "0", "r": "true", "hidden": True}]}
+    IS_ALL_PARAMS_NOT_HIDDEN_INPUTS = [
+        (NO_HIDDEN, True),
+        (HIDDEN_FALSE, True),
+        (HIDDEN_TRUE, False),
+        (HIDDEN_TRUE_AND_FALSE, False),
+    ]
+
+    @pytest.mark.parametrize("current, answer", IS_ALL_PARAMS_NOT_HIDDEN_INPUTS)
+    def test_is_all_params_not_hidden(self, current, answer):
+        structure = mock_structure(current_file=current)
+        validator = IntegrationValidator(structure)
+        assert validator.is_all_params_not_hidden() is answer