+++ CHANGELOG.md
@@ -11,6 +11,7 @@
     * The -o argument is no longer mandatory, default value will be the input file directory.
 * Added support for env var: *DEMISTO_SDK_SKIP_VERSION_CHECK*. When set version checks are skipped.
 * Fixed an issue in which the CHANGELOG files did not match our scheme.
+* Verify there are no hidden integration parameters.
 
 ### 0.4.4
 * Added a validator for IncidentTypes (incidenttype-*.json).